+++ src/zope/schema/interfaces.py
@@ -41,9 +41,12 @@ from zope.schema._bootstrapinterfaces import IContextAwareDefaultFactory
 from zope.schema._bootstrapinterfaces import IFromUnicode
 from zope.schema._bootstrapinterfaces import IValidatable
 from zope.schema._bootstrapinterfaces import InvalidValue
+from zope.schema._bootstrapinterfaces import LenOutOfBounds
 from zope.schema._bootstrapinterfaces import NotAContainer
 from zope.schema._bootstrapinterfaces import NotAnInterface
 from zope.schema._bootstrapinterfaces import NotAnIterator
+from zope.schema._bootstrapinterfaces import OrderableOutOfBounds
+from zope.schema._bootstrapinterfaces import OutOfBounds
 from zope.schema._bootstrapinterfaces import RequiredMissing
 from zope.schema._bootstrapinterfaces import SchemaNotCorrectlyImplemented
 from zope.schema._bootstrapinterfaces import SchemaNotFullyImplemented
@@ -57,6 +60,8 @@ from zope.schema._bootstrapinterfaces import ValidationError
 from zope.schema._bootstrapinterfaces import WrongContainedType
 from zope.schema._bootstrapinterfaces import WrongType
 
+
+
 from zope.schema._compat import PY3
 
 from zope.schema._messageid import _
@@ -68,10 +73,13 @@ __all__ = [
     'InvalidId',
     'InvalidURI',
     'InvalidValue',
+    'LenOutOfBounds',
     'NotAContainer',
     'NotAnInterface',
     'NotAnIterator',
     'NotUnique',
+    'OrderableOutOfBounds',
+    'OutOfBounds',
     'RequiredMissing',
     'SchemaNotCorrectlyImplemented',
     'SchemaNotFullyImplemented',
@@ -85,6 +93,7 @@ __all__ = [
     'ValidationError',
     'WrongContainedType',
     'WrongType',
+
 
     # Interfaces
     'IASCII',